% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_anchors.R
\name{frequency_anchors}
\alias{frequency_anchors}
\title{Generate frequency anchors}
\usage{
frequency_anchors(
  df,
  parsons_col,
  group_id_col,
  individual_id_col,
  call_id_col,
  call_string_col,
  starting_frequency = 4000,
  frequency_shift = 1000
)
}
\arguments{
\item{df}{Data frame. A data frame generated by \code{parsons_code()} that contains the Parsons code strings.}

\item{parsons_col}{Character string. The name of the column in the data frame containing Parsons code strings.}

\item{group_id_col}{Character string. The name of the column in the data frame with numeric group identifiers.}

\item{individual_id_col}{Character string. The name of the column in the data frame containing numeric individual identifiers.}

\item{call_id_col}{Character string. The name of the column in the data frame containing unique numeric identifiers for each vocalization per individual.}

\item{call_string_col}{Character string. The name of the column in the data frame that contains the character string per vocalization.}

\item{starting_frequency}{Numeric value. A numeric value in Hz that specifies the frequency value that will be used as a baseline for creating frequency anchors with Parsons code. For instance, if this value is 4000 Hz and the first Parsons code value is "constant", then the first frequency anchor will be 4000 Hz. The default value is 4000 Hz.}

\item{frequency_shift}{Numeric value. A numeric value in Hz that specifies the frequency value that will be used to shift direction (or not) from the previous frequency anchor based on the Parsons code. For instance, if \code{frequency_shift} is 1000 Hz and the first Parsons code value is "up", then the first frequency anchor will be 5000 Hz. The default value is 1000 Hz, but we have found that when the total string length is over 60 characters, it is better to use a smaller value (100 Hz), which avoids generating negative or zero values.}
}
\value{
A data frame containing the string of the vocalization, the original Parsons code string, and one column per frequency anchor value. This data structure facilitates calculating the minimum and maximum frequency anchors across vocalizations, which can in turn facilitate writing synthetic audio files with \code{soundgen} and making spectrogram image files.
}
\description{
This function converts Parsons code strings to frequency anchors for later guiding frequency modulation patterns in \code{soundgen} audio files.
}
\details{
\code{frequency_anchors()} returns the same data frame that was used as input, and with additional columns that hold frequency values in Hz. These columns will be used as anchors to guide frequency modulation patterns when creating syntheic audio files with the \code{soundgen} package. The starting frequency value is also used to end the frequency anchors. The number of frequency anchor columns in the data frame returned by the function will depend on the length of each string. Currently, this function does not internally correct frequency anchors that are negative or zero, and so the output of this function should be carefully checked prior to creating synthetic files in \code{soundgen}. While testing this function, we found that setting negative or zero values in the resulting data frame to 1000 Hz worked well, but this change has not been thoroughly tested.
}
\examples{
seed <- 8
set.seed(seed) # For reproducibility
library(tidyverse)

example_calls <- generate_strings(n_groups = 2, n_individuals = 5, n_calls = 10, string_length = 16, group_information = 8, individual_information = 2)

example_calls_parsons <- parsons_code(example_calls, "Call", list("A" = "up", "B" = "down", "C" = "constant"))

anchors <- frequency_anchors(example_calls_parsons, "Parsons_Code", "Group", "Individual", "Call_ID", "Call", starting_frequency = 4000, frequency_shift = 1000)

glimpse(anchors)

}
\author{
Ari Cross, Grace Smith-Vidaurre
}
