% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_strings.R
\name{generate_strings}
\alias{generate_strings}
\title{Generate strings with group membership and individual identity information}
\usage{
generate_strings(
  n_groups = 2,
  n_individuals = 5,
  n_calls = 10,
  string_length = 16,
  group_information = 8,
  individual_information = 2
)
}
\arguments{
\item{n_groups}{Integer. The number of groups.}

\item{n_individuals}{Integer. The number of individuals per group.}

\item{n_calls}{Integer. The number of calls per individual.}

\item{string_length}{Integer. The number of characters present in the string that encodes individual-level and group-level information through frequency modulation patterns. The current minimum is 6 characters and the current maximum is 200 characters.}

\item{group_information}{Integer. The number of characters that will vary in the middle of the string across groups. The default is 8 characters.}

\item{individual_information}{Integer. The number of characters that will vary in the middle of the string within groups. The default is 2 characters.}
}
\value{
A data frame containing the call strings as well as metadata columns that contain unique numeric identifiers for the group, individual, and call for the given individual.
}
\description{
\code{generate_strings()} creates character strings that represent vocalizations with varying amounts of group membership and individual identity information. The function facilitates specifying the length of string components that should be allocated to encoding individual identity (characters that are selected to emphasize differentiation among individuals) versus group membership (characters shared among group members but not between groups). Currently, this function builds strings using 3 unique characters only ("A", "B", and "C").
}
\details{
The individual-specific and group-specific string components are combined to form the middle of a longer string. The individually-specific component of the string may not be unique to any one individual in a group, as individual distinctiveness will depend on the total number of individuals in a group, the length of the individually-specific string component, and depending on how users modify this function, the number of unique characters or symbols that can be used to create strings. For instance, when the length of the individually distinctive string component is 2 characters long, and you're using 3 unique characters to create strings, there will be 3^2 or 9 possible unique individual signatures.

The final string is composed of a global head (a short string of characters shared across all individuals), the group membership information, individual identity information, and a global tail (a short string of characters shared across all individuals). The global heads and tails are used to guide the start and end of frequency modulation patterns created after converting character strings to Parsons code in later functions. The relative amount of group versus individual information across calls can be controlled by setting the length of \code{group_information} and \code{individual_information}, respectively. For example, when \code{group_information} is longer than \code{individual_information}, there will be more group membership information encoded in strings, and vice versa. The current version of the function does not facilitate varying string length within or across individuals.
}
\examples{
seed <- 8
set.seed(seed) # For reproducibility
library(tidyverse)
example_calls <- generate_strings(n_groups = 2, n_individuals = 5, n_calls = 10, string_length = 16, group_information = 8, individual_information = 2)
glimpse(example_calls)

}
\author{
Ari Cross, Grace Smith-Vidaurre
}
