% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsons_code.R
\name{parsons_code}
\alias{parsons_code}
\title{Convert character strings to Parsons code}
\usage{
parsons_code(
  df,
  string_col,
  global_head_col,
  group_head_col,
  individual_middle_col,
  random_variation_col,
  group_tail_col,
  global_tail_col,
  mapping = list(A = "up", B = "down", C = "constant")
)
}
\arguments{
\item{df}{Data frame. A data frame object produced by \code{generate_strings()}. Each row is a string of characters representing a vocalization.}

\item{string_col}{Character string. The name of the column in \code{df} that contains the strings.}

\item{global_head_col}{Character string. The name of the column in \code{df} that contains the global head information.}

\item{group_head_col}{Character string. The name of the column in \code{df} that contains the group head information.}

\item{individual_middle_col}{Character string. The name of the column in \code{df} that contains the individual middle information.}

\item{random_variation_col}{Character string. The name of the column in \code{df} that contains the random variation information.}

\item{group_tail_col}{Character string. The name of the column in \code{df} that contains the group tail information.}

\item{global_tail_col}{Character string. The name of the column in \code{df} that contains the global tail information.}

\item{mapping}{List. A list for Parsons code mapping. Each list element contains each character value and its corresponding directionality in Parsons code. The default mapping are the first three letters of the alphabet in uppercase. This creates a mapping for 1-base Parsons code (three unique directions: up, down, and constant).}
}
\value{
This function returns the same data frame that was used as input for the function with an additional column of the string converted to Parsons code.
}
\description{
This function converts the character strings generated by \code{generate_strings()} into Parsons code.
}
\details{
This function converts the character strings generated by \code{generate_strings()} into Parsons code. The current version of this function was written for 1-base Parsons code and has not been tested for finer-grained Parsons code. The assignment of character values to each Parsons code value is subjective, and changing this mapping will change how frequency modulation patterns are generated downstream.
}
\examples{
seed <- 8
set.seed(seed) # For reproducibility
library(tidyverse)

example_calls <- generate_strings(n_groups = 2,
                                   n_individuals = 5,
                                   n_calls = 10,
                                   string_length = 16,
                                   group_information = 8,
                                   individual_information = 2,
                                   random_variation = 2)

example_calls_parsons <- parsons_code(example_calls,
                                      "Call",
                                      "Global_head",
                                      "Group_head",
                                      "Individual_middle",
                                      "Random_variation",
                                      "Group_tail",
                                      "Global_tail",
                                      list("A" = "up",
                                           "B" = "down",
                                           "C" = "constant")
                                )
glimpse(example_calls_parsons)

}
\author{
Ari Cross, Grace Smith-Vidaurre, Raneem Samman
}
