% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsons_code.R
\name{parsons_code}
\alias{parsons_code}
\title{Convert character strings to Parsons code}
\usage{
parsons_code(
  df,
  string_col,
  mapping = list(A = "up", B = "down", C = "constant")
)
}
\arguments{
\item{df}{Data frame. A data frame object produced by \code{generate_strings()}. Each row is a string of characters representing a vocalization.}

\item{string_col}{Character string. The name of the column in \code{df} that contains the strings.}

\item{mapping}{List. A list for Parsons code mapping. Each list element contains the each character value and its corresponding directionality in Parsons code. The default mapping are the first three letters of the alphabet in uppercase, to create a  mapping for 1-base Parsons code (three unique directions: up, down, and constant).}
}
\value{
The same data frame that was used as input for the function, with an additional column of the string converted to Parsons code.
}
\description{
This function generates Parsons code from the character strings produced by \code{generate_strings()}.
}
\details{
This function converts the character strings generated by \code{generate_strings()} to Parsons code. The current version of this function was written for 1-base Parsons code and has not been tested for finer-grained Parsons code. The assignment of character values to each Parsons code value is subjective, and changing this mapping will change how frequency modulation patterns are generated downstream.
}
\examples{
seed <- 8
set.seed(seed) # For reproducibility
library(tidyverse)

example_calls <- generate_strings(n_groups = 2, n_individuals = 5, n_calls = 10, string_length = 16, group_information = 8, individual_information = 2)

example_calls_parsons <- parsons_code(example_calls, "Call", list("A" = "up", "B" = "down", "C" = "constant"))
glimpse(example_calls_parsons)

}
\author{
Ari Cross, Grace Smith-Vidaurre
}
